+++ Craftsman/Models/ApiTemplate.cs
@@ -3,6 +3,8 @@ using System.Linq;
 
 namespace Craftsman.Models
 {
+    using System.Net;
+    using System.Net.Sockets;
     using Humanizer;
 
     /// <summary>
@@ -88,17 +90,6 @@ namespace Craftsman.Models
 
         public bool UseSoftDelete { get; set; } = true;
 
-        private DockerConfig _dockerConfig;
-        public DockerConfig DockerConfig
-        {
-            get
-            {
-                var projectNamelessConfig = _dockerConfig ?? new DockerConfig();
-                projectNamelessConfig.ProjectName = ProjectName;
-                projectNamelessConfig.Provider = DbContext.Provider;
-                return projectNamelessConfig;
-            }
-            set => _dockerConfig = value;
-        }
+        public DockerConfig DockerConfig { get; set; } = new DockerConfig();
     }
 }